personel profile

+++ Controllers/Account.cs
@@ -14,6 +14,12 @@ public class CreateUserModel
     public string password { get; set; } = string.Empty;
     public string role { get; set; } = string.Empty;
     public string CompanyName { get; set; } = string.Empty;
+    public string FirstName { get; set; } = string.Empty;
+    public string LastName { get; set; } = string.Empty;
+    public DateTime BirthDate { get; set; }
+    public string Department { get; set; } = string.Empty;
+    public List<DateTime> LeaveDay { get; set; } = new List<DateTime>();
+    public Dictionary<string, string> CVInfos { get; set; } = new Dictionary<string, string>();
 }
 
 public class LoginModel
@@ -43,21 +49,33 @@ public class AccountController : Controller
     //public async Task<IActionResult> CreateUser(string username, string email, string password, string role)
     public async Task<IActionResult> CreateUser([FromBody] CreateUserModel model)
     {
-        var user = new ApplicationUser { UserName = model.username, Email = model.email, CompanyName = model.CompanyName};
+        var existingUser = await _userManager.FindByNameAsync(model.username);
+        if (existingUser != null)
+        {
+            return BadRequest(new { message = "Username already has ben taken " });
+        }
+        var user = new ApplicationUser { UserName = model.username, 
+            Email = model.email, 
+            CompanyName = model.CompanyName,
+            FirstName = model.FirstName,
+            LastName = model.LastName,
+            BirthDate = model.BirthDate,
+            //CVInfos = model.CVInfos.Select(kv => new CVInfo { Key = kv.Key, Value = kv.Value }).ToList()
+        };
         var result = await _userManager.CreateAsync(user, model.password);
         if (result.Succeeded)
         {
             if (!await _roleManager.RoleExistsAsync(model.role))
             {
                 var roleeror = await _userManager.DeleteAsync(user);
-                return BadRequest("Role bulunamadi."+ " delete error: "+ roleeror.ToString());
+                return BadRequest(new { message = "Role bulunamadi."+ " delete error: "+ roleeror.ToString() });
             }
             await _userManager.AddToRoleAsync(user, model.role);
             return Ok("Kullanici basariyla oluşturuldu ve rol atandi.");
         }
         var resultErrors = await _userManager.DeleteAsync(user);
 
-        return BadRequest("Kullanici oluşturulurken hata oluştu. "+ result.ToString()+ " delete error: "+ resultErrors.ToString());
+        return BadRequest(new { message = "Kullanici oluşturulurken hata oluştu. "+ result.ToString()+ " delete error: "+ resultErrors.ToString() });
     }
 
 /*     [HttpGet("getuser")]
